requirements.txt
aiohttp
pandas
numpy
pyyaml
google-generativeai
fastapi
uvicorn
websockets
python-dotenv
requests
ccxt
